refactor(dispatcher): 메시지 디스패칭 로직 리팩토링 및 타입 안정성 강화
- IMessage::type 타입을 uint32_t에서 MessageType 열거형으로 변경
- switch-case 및 메시지 생성자에서의 불필요한 캐스팅 제거
- Process() 로직을 각 메시지 타입별 핸들러 함수로 분리하여 가독성 개선
- DispatcherImpl의 미처리 메시지 타입에 대한 LOG_INFO 추가

+++ src/System/Dispatcher/SystemMessages.h
@@ -14,7 +14,7 @@ struct TimerMessage : public IMessage
 {
     TimerMessage()
     {
-        type = (uint32_t)MessageType::LOGIC_TIMER;
+        type = MessageType::LOGIC_TIMER;
     }
     uint32_t timerId;
     void *pParam;
@@ -26,7 +26,7 @@ struct TimerExpiredMessage : public IMessage
 {
     TimerExpiredMessage()
     {
-        type = (uint32_t)MessageType::LOGIC_TIMER_EXPIRED;
+        type = MessageType::LOGIC_TIMER_EXPIRED;
     }
     uint64_t timerId;
 };
@@ -35,7 +35,7 @@ struct TimerTickMessage : public IMessage
 {
     TimerTickMessage()
     {
-        type = (uint32_t)MessageType::LOGIC_TIMER_TICK;
+        type = MessageType::LOGIC_TIMER_TICK;
     }
     uint32_t tickCount;
 };
@@ -44,7 +44,7 @@ struct TimerAddMessage : public IMessage
 {
     TimerAddMessage()
     {
-        type = (uint32_t)MessageType::LOGIC_TIMER_ADD;
+        type = MessageType::LOGIC_TIMER_ADD;
     }
     uint64_t timerId;
     uint32_t logicTimerId;
@@ -59,7 +59,7 @@ struct TimerCancelMessage : public IMessage
 {
     TimerCancelMessage()
     {
-        type = (uint32_t)MessageType::LOGIC_TIMER_CANCEL;
+        type = MessageType::LOGIC_TIMER_CANCEL;
     }
     uint64_t timerId;
     ITimerListener *listener;
